+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - "View previous version" button on questions whose answer changed since the previous report; opens a modal showing the previous answer
-- Sysadmin commands to export and atomically import incident workflow configuration for a sector regulation (#826)
+- Sysadmin commands to export and atomically import incident workflow configuration, including missing sectors, for a sector regulation (#826)
 
 ### Fixed